Conflicts:
	roads.py

+++ tools.py
@@ -241,7 +241,7 @@ class Probe_fgelev(object):
                 return -100
 
             try:
-                self.fgelev_pipe.stdin.write("%i %g %g\r\n" % (self.record, position.lon, position.lat))
+                self.fgelev_pipe.stdin.write("%i %1.10f %1.10f\r\n" % (self.record, position.lon, position.lat))
             except IOError, reason:
                 logging.error(reason)
 
@@ -762,8 +762,10 @@ def init(new_transform):
 def install_files(file_list, dst):
     """link files in file_list to dst"""
     for the_file in file_list:
-        the_dst = dst + os.sep + the_file
+        the_dst = dst # + os.sep + the_file
         logging.info("cp %s %s" % (the_file, the_dst))
+        if os.path.exists(the_dst + the_file):
+            return
         try:
             shutil.copy2(the_file, the_dst)
         except OSError, reason:
